+++ lib/plugin/pl_player/index.dart
@@ -15,3 +15,4 @@ export './utils.dart';
 export './engine/player_engine_interface.dart';
 export './engine/impl/mpv_player_engine.dart';
 export './engine/impl/headless_player_engine.dart';
+export './engine/impl/media3_player_engine.dart';